Changes in the memory retrieval and update memory_manger prompt and create a intial workflow

+++ memory/episodic.py
@@ -2,7 +2,7 @@ import sqlite3
 from pathlib import Path
 import uuid
 
-from graph.state import MarvelState
+from graph.state import MemoryManagerOutput
 BASE_DIR = Path(__file__).resolve().parent.parent
 DB_PATH = BASE_DIR / "database" / "episodic.db"
 
@@ -17,7 +17,7 @@ def init_db():
 
     cursor.execute("""
         CREATE TABLE IF NOT EXISTS episodic_memories (
-        memory_id INTEGER PRIMARY KEY AUTOINCREMENT,
+        memory_id INTEGER PRIMARY KEY,
         user_id TEXT NOT NULL,
         event TEXT NOT NULL,
         date TEXT,
@@ -34,20 +34,20 @@ def init_db():
     conn.commit()
     conn.close()
 
-def create_episodic_memory(user_id, event, date, context, summary, importance, confidence, source, created_at):
+def create_episodic_memory(memory_id, user_id, event, date, context, summary, importance, confidence, source):
     conn = get_connection()
     cursor = conn.cursor()
 
 
     cursor.execute("""
-        INSERT INTO episodic_memories (user_id, event, date, context, summary, importance, confidence, source, created_at, updated_at)
-        VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-    """, (user_id, event, date, context, summary, importance, confidence, source, created_at, created_at))
+        INSERT INTO episodic_memories (memory_id, user_id, event, date, context, summary, importance, confidence, source, created_at, updated_at)
+        VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)
+    """, (memory_id, user_id, event, date, context, summary, importance, confidence, source))
 
-    memory_id = cursor.lastrowid
+    
     conn.commit()
     conn.close()
-    return memory_id
+
 
 def update_episodic_memory(memory_id, user_id, event, date, context, summary, importance, confidence, source , updated_at):
     conn = get_connection()
@@ -66,7 +66,7 @@ def update_episodic_memory(memory_id, user_id, event, date, context, summary, im
             updated_at = CURRENT_TIMESTAMP
         WHERE memory_id = ?
         AND user_id = ?
-    """, (event, date, context, summary, importance, confidence, source, updated_at))
+    """, (event, date, context, summary, importance, confidence, source, memory_id, user_id))
 
     conn.commit()
     updated = cursor.rowcount
@@ -89,29 +89,35 @@ def delete_episodic_memory(memory_id, user_id):
     return deleted > 0  # Return True if a row was deleted, False otherwise
 
 
-def retrieve_episodic_memory(date: str | None, user_id: str):
+def retrieve_episodic_memory(date: str | None, user_id: str, memory_id:list):
 
     conn = get_connection()
     conn.row_factory = sqlite3.Row
     cursor = conn.cursor()
+    memories = []
+    
+    for i in memory_id:
+        if date is None:
+            cursor.execute("""
+                SELECT *
+                FROM episodic_memories
+                WHERE user_id = ?
+                AND memory_id = ?
+                ORDER BY date DESC
+            """, (user_id, i))
+        else:
+            cursor.execute("""
+                SELECT *
+                FROM episodic_memories
+                WHERE date = ?
+                AND user_id = ?
+                AND memory_id = ?
+                ORDER BY date DESC
+            """, (date, user_id, i))
+        row = cursor.fetchall()
+        memories.append(dict(row))
 
-    if date is None:
-        cursor.execute("""
-            SELECT *
-            FROM episodic_memories
-            WHERE user_id = ?
-            ORDER BY date DESC
-        """, (user_id,))
-    else:
-        cursor.execute("""
-            SELECT *
-            FROM episodic_memories
-            WHERE date = ?
-            AND user_id = ?
-            ORDER BY date DESC
-        """, (date, user_id))
-
-    memories = [dict(row) for row in cursor.fetchall()]
+    
 
     conn.close()
 
@@ -120,35 +126,45 @@ def retrieve_episodic_memory(date: str | None, user_id: str):
 
 
 
-def handle_episodic_memory(state: MarvelState, memory: dict):
+def handle_episodic_memory(state):
+    """Handles all the CRUD operation for episodic memeory."""
 
-    user_id = state["user_id"]
-    action = memory["action"]
-    confidence = memory.get("confidence", 1.0)
-    importance = memory.get("importance", 1.0)
-    source = memory.get("source", "unknown")
-    data = memory.get("data", {})
-    reason = memory.get("reason", "")
+    i=0
+    for operation in state.get("memories", []):
+        if operation.get("memory_type") != "episodic":
+            continue
+            
+        action = operation.get("action")
+        confidence = operation.get("confidence", 1.0)
+        source = operation.get("source", "unknown")
+        data = operation.get("data", {})
+        importance = data.get("importance", 1.0)
+        user_id = operation.get("user_id")
+        memory_id = operation.get("memory_id")[i]
 
+        if not user_id:
+            return {"status": "error", "message": "user_id is missing from state"}
 
-    if action == "create":
-        memory_id = create_episodic_memory(user_id=user_id, event=data.get("event"),date=data.get("date"), context=data.get("context"), summary=data.get("summary"), importance=importance, confidence=confidence, source=source, created_at=data.get("date"))
-        return {"status": "success", "memory_id": memory_id}
-    elif action == "update":
-        memory_id = memory.get("memory_id")
-
-        if not memory_id:
-            return {
-                "status": "error",
-                "message": "memory_id required for update"
-            }
-        updated = update_episodic_memory(memory_id=memory_id, user_id=user_id, event=data.get("event"),date=data.get("date"), context=data.get("context"), summary=data.get("summary"), importance=importance, confidence=confidence, source=source, updated_at=data.get("date"))
-        return {"status": "success" if updated else "not_found"}
-    elif action == "delete":
-        deleted = delete_episodic_memory(memory_id=memory_id, user_id=user_id)
-        return {"status": "success" if deleted else "not_found"}
-                    
-    elif action == "retrieve":
-        date = data.get("date")
-        memories = retrieve_episodic_memory(date=date, user_id=user_id)
-        return {"status": "success", "memories": memories}
+        if action == "create":
+            create_episodic_memory(memory_id=memory_id, user_id=user_id, event=data.get("event"),date=data.get("date"), context=data.get("context"), summary=data.get("summary"), importance=importance, confidence=confidence, source=source)
+            print("episodic create success")
+        elif action == "update":
+            if not memory_id:
+                return {
+                    "status": "error",
+                    "message": "memory_id required for update"
+                }
+            update_episodic_memory(memory_id=memory_id, user_id=user_id, event=data.get("event"),date=data.get("date"), context=data.get("context"), summary=data.get("summary"), importance=importance, confidence=confidence, source=source, updated_at=data.get("date"))
+            print("episodic update success")
+        elif action == "delete":
+            delete_episodic_memory(memory_id=memory_id, user_id=user_id)
+            #return {"status": "success" if deleted else "not_found"}
+            print("episodic delete success")
+                        
+        elif action == "retrieve":
+            date = data.get("date")
+            memories = retrieve_episodic_memory(date=date, user_id=user_id, memory_id= memory_id)
+            return {"memories": memories}
+        else:
+            print("error in episodoic.py")
+        i=i+1
